requirements.txt
google-generativeai
pdfplumber
playwright
pytest
doctype
json
langgraph
bs4
